+++ utils/userEditForm.ts
@@ -2,7 +2,6 @@ import { User, UserEditForm } from 'types/user';
 
 export const userEditForm = (user: User): UserEditForm => {
   return {
-    email: user.email,
     nickname: user.nickname,
     description: user.description,
     profileImage: user.profileImage,
